requirements.txt
crewai
load-dotenv
crewai-tools
langchain-groq
pydantic
